extract virtual_env method to own module
