+++ pure/prompt.py
@@ -1,7 +1,7 @@
 import argparse
 import os
 
-from pure import colors, repository
+from pure import colors, repository, virtual_env
 
 SUCCESS = 0
 
@@ -15,13 +15,6 @@ def current_working_path():
     return colors.info(os.getcwd())
 
 
-def virtual_env():
-    if 'VIRTUAL_ENV' in os.environ:
-        virtual_env_name = os.path.basename(os.environ['VIRTUAL_ENV'])
-        return colors.mute('{} '.format(virtual_env_name))
-    return ''
-
-
 def layout():
     return "\n{current_working_path} {git_active_branch}{git_is_dirty}\n{virtual_env}{prompt_symbol} "
 
@@ -31,7 +24,7 @@ def prompt(args):
         'current_working_path': current_working_path(),
         'git_active_branch': repository.active_branch(os.getcwd()),
         'git_is_dirty': repository.is_dirty(os.getcwd()),
-        'virtual_env': virtual_env(),
+        'virtual_env': virtual_env.name(),
         'prompt_symbol': prompt_symbol(args.last_command_status)
     }
     print(layout().format(**data), end='')
